third_party/blink/web_tests/compositing/squashing/dont-squash-into-videos.html

<video id="video" style="will-change: transform; transform: translateZ(0); position: absolute; top: 0px; left: 0px; height: 100px; width: 100px; background: lightblue"></video>
<div id="target" style="position: absolute; top: 50px; left: 50px; width: 200px; height: 200px; background: lightgray;"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
onload = function() {
    test(function() {
        if (window.internals) {
            var layer_tree = internals.layerTreeAsText(document);
            assert_true(layer_tree.includes("id='video'"), 'composited video');
            assert_true(layer_tree.includes("id='target'"), 'should not be squashed into video');
        }
    });
};
</script>
